refactor: delete console.log with split funciton
