+++ src/components/ranking/SelectSeason.client.tsx
@@ -14,6 +14,10 @@ const SelectSeason = ({ selectedSeasonId }: { selectedSeasonId: string }) => {
     queryFn: getSeason,
   });
 
+  const selectedSeasonIndex = seasonList
+    .map((v) => v.id)
+    .indexOf(Number(selectedSeasonId));
+
   if (isLoading) {
     return <Loading />;
   }
@@ -24,7 +28,9 @@ const SelectSeason = ({ selectedSeasonId }: { selectedSeasonId: string }) => {
         <Menu.Button className="inline-flex h-[75px] w-full items-center justify-center rounded-[20px] border-[1px] border-[#D9D9D9] bg-white px-4 py-2 text-lg font-medium text-[#828282] shadow-md hover:bg-[#e6e6e6] focus:outline-none focus-visible:ring-2 focus-visible:ring-white focus-visible:ring-opacity-75">
           <ChevronDownIcon className="m-auto mr-2 h-5 w-5" aria-hidden="true" />
           <div className="w-[160px] rounded-[20px] bg-[#F9F6ED] p-2 text-[25px] font-bold">
-            {selectedSeasonId}
+            {selectedSeasonIndex >= 0
+              ? seasonList[selectedSeasonIndex].description
+              : ""}
           </div>
         </Menu.Button>
         <Transition
@@ -49,10 +55,7 @@ const SelectSeason = ({ selectedSeasonId }: { selectedSeasonId: string }) => {
                               ? "bg-violet-500 text-white"
                               : "text-gray-900"
                           } group flex w-full items-center rounded-md px-2 py-2 text-sm`}
-                          // onClick={() => setSelectedSeason(v)}
                         >
-                          <span>{v.start_date}</span>
-                          <span>{v.end_date}</span>
                           <span>{v.description}</span>
                         </button>
                       </Link>
